+++ lightweight-build/public/main.css
@@ -1,7 +1,7 @@
 
 body {
     display: flex;
-    background-image: linear-gradient(to top left, #f598a8, #f6edb2);
+    background-image: linear-gradient(to top right, rgb(198, 198, 245) , rgb(255, 217, 236));
     background-size: 200% 4000px;
     animation: gradient 10s ease infinite;
     height: 100vh;
