+++ lightweight-build/public/main.js
@@ -1,308 +1,1 @@
-const svg = d3.select("#svg");
-const { width, height } = document
-  .getElementById("svg")
-  .getBoundingClientRect();
-
-
-var type = 2; // 0: mood, 1: sleepminutes, 2: activeminutes, 4: etc.
-let globalData; // Declare a global variable to store the parsed data
-function tooltip(d) {
-  const formatDate = d3.timeFormat("%B %d, %Y");
-  switch (type) {
-    case 0:
-      return `${formatDate(d.date)}: mood of ${d.value.toFixed(0)}`
-    case 1:
-      return `${formatDate(d.date)}: ${d.value.toFixed(0)} minutes of sleep`
-    case 2:
-      return `${formatDate(d.date)}: ${d.value.toFixed(0)} minutes of activity`
-    default:
-      return "something weng wrong"
-  }
-}
-function drawCalendar() {
-  var dateValues;
-  switch (type) {
-    case 0:
-      dateValues = globalData.map(dv => ({
-        date: d3.utcDay(new Date(dv.Date)),
-        value: Number(dv.Mood)
-      }));
-      break;
-    case 1:
-      dateValues = globalData.map(dv => ({
-        date: d3.utcDay(new Date(dv.Date)),
-        value: Number(dv.SleepMinutes)
-      }));
-      break;
-    case 2:
-      dateValues = globalData.map(dv => ({
-        date: d3.utcDay(new Date(dv.Date)),
-        value: Number(dv.ActiveMinutes),
-        
-      }));
-      break;
-    default:
-      console.log("uhhhh");
-
-  }
-  console.log(dateValues)
-
-  const categorical = type == 0;
-  const years = Array.from(
-    d3.group(dateValues, d => d.date.getUTCFullYear()),
-    ([key, values]) => ({ key, values })
-  ).reverse();
-  const values = dateValues.map(c => c.value);
-  const maxValue = d3.max(values);
-  const minValue = d3.min(values);
-
-  const cellSize = 17;
-  const yearHeight = cellSize * 7;
-
-  var group = svg.select("g");
-  // If the <g> element doesn't exist yet, append it
-  if (group.empty()) {
-      group = svg.append("g");
-  }
-
-  const year = group
-    .selectAll("g")
-    .data(years)
-    .join("g")
-    .attr(
-      "transform",
-      (d, i) => `translate(50, ${yearHeight * i + cellSize * 1})`
-    );
-  year
-    .append("text")
-    .attr("x", -40)
-    .attr("y", -35)
-    .attr("text-anchor", "end")
-    .attr("font-size", 16)
-    .attr("font-weight", 550)
-    .attr("transform", "rotate(270)")
-    .text(d => d.key);
-
-  const formatDay = d =>
-    ["Su", "Mo", "Tu", "We", "Th", "Fr", "Sa"][d.getUTCDay()];
-  const countDay = d => d.getUTCDay();
-  const timeWeek = d3.utcSunday;
-  var colorFn = d3
-    .scaleSequential(d3.interpolateBuGn)
-    .domain([Math.floor(minValue), Math.ceil(maxValue)]);
-  if (categorical) {
-    colorFn = d3.scaleQuantize()
-      .domain([1, 5])
-      .range(d3.schemeBuGn[5]);
-  }
-  year
-    .append("g")
-    .attr("text-anchor", "end")
-    .selectAll("text")
-    .data(d3.range(7).map(i => new Date(1995, 0, i)))
-    .join("text")
-    .attr("x", -5)
-    .attr("y", d => (countDay(d) + 0.5) * cellSize)
-    .attr("dy", "0.31em")
-    .attr("font-size", 15)
-    .text(formatDay);
-  year
-    .append("g")
-    .selectAll("rect")
-    .data(d => d.values)
-    .join("rect")
-    .attr("width", cellSize - 1.5)
-    .attr("height", cellSize - 1.5)
-    .attr(
-      "x",
-      (d, i) => timeWeek.count(d3.utcYear(d.date), d.date) * cellSize + 10
-    )
-    .attr("y", d => countDay(d.date) * cellSize + 0.5)
-    .attr("fill", d => colorFn(d.value))
-    .on("mouseover", (event, d) => {
-      // Get mouse position
-      const [x, y] = d3.pointer(event);
-      // Show tooltip
-      d3.select("body").append("div")
-          .attr("class", "tooltip")
-          .style("position", "absolute")
-          .style("left", `${x}px`)
-          .style("top", `${y + 190}px`)
-          .html(`${tooltip(d)}`);
-      })
-      .on("mouseout", () => {
-          // Hide tooltip on mouseout
-          d3.select(".tooltip").remove();
-      });
-
-  const legend = group
-    .append("g")
-    .attr(
-      "transform",
-      `translate(10, ${years.length * yearHeight + cellSize * 4})`
-    );
-
-  const categoriesCount = 10;
-  var categories = [...Array(categoriesCount)].map((_, i) => {
-    const upperBound = (maxValue / categoriesCount) * (i + 1);
-    const lowerBound = (maxValue / categoriesCount) * i;
-    return {
-        upperBound,
-        lowerBound,
-        color: d3.interpolateBuGn(upperBound / maxValue),
-        selected: true
-    };
-  }); 
-
-  if (categorical) {
-    categories = [{upperBound: 1, lowerBound: 0, color: colorFn(1), selected: true}, 
-                  {upperBound: 2, lowerBound: 1, color: colorFn(2), selected: true}, 
-                  {upperBound: 3, lowerBound: 2, color: colorFn(3), selected: true}, 
-                  {upperBound: 4, lowerBound: 3, color: colorFn(4), selected: true}, 
-                  {upperBound: 5, lowerBound: 4, color: colorFn(5), selected: true}];
-  }
-
-  const legendWidth = 60;
-
-  function toggle(legend) {
-    const { lowerBound, upperBound, selected } = legend.explicitOriginalTarget.__data__;
-    console.log(legend);
-    console.log(lowerBound, upperBound, selected);
-
-    legend.explicitOriginalTarget.__data__.selected = !selected;
-
-    const highlightedDates = years.map(y => ({
-      key: y.key,
-      values: y.values.filter(
-      v => v.value > lowerBound && v.value <= upperBound
-      )
-    }));
-
-    year
-      .data(highlightedDates)
-      .selectAll("rect")
-      .data(d => d.values, d => d.date)
-      .transition()
-      .duration(500)
-      .attr("fill", d => (legend.explicitOriginalTarget.__data__.selected ? colorFn(d.value) : "white"));
-  }
-
-  console.log(categories);
-  if (categorical) {
-  legend
-    .selectAll("rect")
-    .data(categories)
-    .enter()
-    .append("rect")
-    .attr("fill", d => d.color)
-    .attr("x", (d, i) => legendWidth * i)
-    .attr("width", legendWidth)
-    .attr("height", 15)
-    .on("click", toggle);
-
-  legend
-    .selectAll("text")
-    .data(categories)
-    .join("text")
-    .attr("transform", "rotate(90)")
-    .attr("y", (d, i) => -legendWidth * i)
-    .attr("dy", -30)
-    .attr("x", 18)
-    .attr("text-anchor", "start")
-    .attr("font-size", 11)
-    .text(d => `${d.upperBound}`);
-  } else {
-    legend
-        .selectAll("rect")
-        .data(categories)
-        .enter()
-        .append("rect")
-        .attr("fill", d => d.color)
-        .attr("x", (d, i) => legendWidth * i)
-        .attr("width", legendWidth)
-        .attr("height", 15)
-        .on("click", toggle);
-
-    legend
-        .selectAll("text")
-        .data(categories)
-        .join("text")
-        .attr("x", (d, i) => legendWidth * i + 10)
-        .attr("y", -5)
-        .attr("text-anchor", "start")
-        .attr("font-size", 11)
-        .text(d => `${d.lowerBound.toFixed(0)} - ${d.upperBound.toFixed(0)}`);
-  }
-
-}
-
-
-
-
-document.getElementById('uploadForm').addEventListener('submit', function(event) {
-  event.preventDefault(); // Prevent the form from submitting
-  const fileInput = document.getElementById('myfile');
-  const file = fileInput.files[0]; // Get the uploaded file
-  if (file) {
-    const reader = new FileReader();
-    reader.onload = function(e) {
-      const contents = e.target.result; // Get the contents of the file
-      processData(contents); // Call a function to process the data
-    };
-
-    reader.readAsText(file); // Read the file as text
-  }
-});
-
-function processData(contents) {
-  // Here you can parse and process the contents of the file
-  // For example, you can split it by lines
-  console.log("Process");
-
-  // Assign the parsed data to the global variable
-  globalData = d3.csvParse(contents);
-  changeButton();
-  drawCalendar();
-}
-
-
-document.addEventListener("DOMContentLoaded", function() {
-  // Get references to the arrow buttons
-  const leftArrow = document.querySelector('.arrow-left');
-  const rightArrow = document.querySelector('.arrow-right');
-
-  // Add click event listeners to the arrow buttons
-  leftArrow.addEventListener('click', function() {
-    // Action to perform when left arrow button is clicked
-    console.log('Left arrow button clicked');
-    // Add your custom logic here
-    type = (type + 2) % 3;
-    changeButton();
-    drawCalendar();
-  });
-
-  rightArrow.addEventListener('click', function() {
-    // Action to perform when right arrow button is clicked
-    console.log('Right arrow button clicked');
-    type = (type + 1) % 3;
-    changeButton();
-    drawCalendar();
-  });
-});
-
-function changeButton() {
-  const buttonText = document.getElementById('button-text');
-  switch (type) {
-    case 0: 
-      buttonText.textContent = "Mood Heatmap"
-      break;
-    case 1: 
-      buttonText.textContent = "Minutes of Sleep Heatmap"
-      break;
-    case 2: 
-      buttonText.textContent = "Minutes of Activity Heatmap"
-      break;
-    default:
-      console.log("UHHHH");
-  }
-}
+(()=>{const t=d3.select("#svg"),{width:e,height:a}=document.getElementById("svg").getBoundingClientRect(),o={1:"very negative",2:"slightly negative",3:"neutral",4:"slightly positive",5:"very positive"};var n=0;let r;function l(){var e;switch(n){case 0:e=r.map((t=>({date:d3.utcDay(new Date(t.Date)),value:Number(t.Mood)})));break;case 1:e=r.map((t=>({date:d3.utcDay(new Date(t.Date)),value:Number(t.SleepMinutes)})));break;case 2:e=r.map((t=>({date:d3.utcDay(new Date(t.Date)),value:Number(t.ActiveMinutes)})));break;default:console.log("uhhhh")}console.log(e);const a=0==n,l=Array.from(d3.group(e,(t=>t.date.getUTCFullYear())),(([t,e])=>({key:t,values:e}))).reverse(),d=e.map((t=>t.value)),c=d3.max(d),s=d3.min(d),i=c-s;var u=t.select("g");u.empty()&&(u=t.append("g"));const p=u.selectAll("g").data(l).join("g").attr("transform",((t,e)=>`translate(50, ${119*e+17})`));p.append("text").attr("x",-40).attr("y",-35).attr("text-anchor","end").attr("font-size",16).attr("font-weight",550).attr("transform","rotate(270)").text((t=>t.key));const g=t=>t.getUTCDay(),m=d3.utcSunday;var v=d3.scaleSequential(d3.interpolatePlasma).domain([Math.floor(s),Math.ceil(c)]);a&&(v=d3.scaleQuantize().domain([1,5]).range(d3.range(5).map((function(t){return d3.interpolatePlasma(t/4)})))),p.append("g").attr("text-anchor","end").selectAll("text").data(d3.range(7).map((t=>new Date(1995,0,t)))).join("text").attr("x",-5).attr("y",(t=>17*(g(t)+.5))).attr("dy","0.31em").attr("font-size",15).text((t=>["Su","Mo","Tu","We","Th","Fr","Sa"][t.getUTCDay()])),p.append("g").selectAll("rect").data((t=>t.values)).join("rect").attr("width",15.5).attr("height",15.5).attr("x",((t,e)=>17*m.count(d3.utcYear(t.date),t.date)+10)).attr("y",(t=>17*g(t.date)+.5)).attr("fill",(t=>v(t.value))).on("mouseover",((t,e)=>{const[a,r]=d3.pointer(t);console.log(a,r),d3.select("body").append("div").attr("class","tooltip").style("position","absolute").style("left",`${a+35}px`).style("top",`${r+90}px`).html(`${function(t){const e=d3.timeFormat("%B %d, %Y");switch(n){case 0:return`You had a ${o[t.value]} mood on ${e(t.date)}}`;case 1:return`${e(t.date)}: ${t.value.toFixed(0)} minutes of sleep`;case 2:return`${e(t.date)}: ${t.value.toFixed(0)} minutes of activity`;default:return"something went wrong"}}(e)}`)})).on("mouseout",(()=>{d3.select(".tooltip").remove()}));const f=u.append("g").attr("transform",`translate(10, ${119*l.length+68})`);var x=[...Array(5)].map(((t,e)=>{const a=s+i/5*(e+1),o=s+i/5*e;return{upperBound:a,lowerBound:o,color:v((a+o)/2),selected:!0}}));a&&(x=[{upperBound:1,lowerBound:0,color:v(1),selected:!0},{upperBound:2,lowerBound:1,color:v(2),selected:!0},{upperBound:3,lowerBound:2,color:v(3),selected:!0},{upperBound:4,lowerBound:3,color:v(4),selected:!0},{upperBound:5,lowerBound:4,color:v(5),selected:!0}]);const y=60,h=500;function w(t){const{lowerBound:e,upperBound:a,selected:o}=t.explicitOriginalTarget.__data__;console.log(t),console.log(e,a,o),t.explicitOriginalTarget.__data__.selected=!o;const n=l.map((t=>({key:t.key,values:t.values.filter((t=>t.value>e&&t.value<=a))})));p.data(n).selectAll("rect").data((t=>t.values),(t=>t.date)).transition().duration(500).attr("fill",(e=>t.explicitOriginalTarget.__data__.selected?v(e.value):"white"))}console.log(x),a?(f.selectAll("rect").data(x).enter().append("rect").attr("fill",(t=>t.color)).attr("x",((t,e)=>y*e+h)).attr("width",y).attr("height",15).on("click",w),f.selectAll("text").data(x).join("text").attr("x",((t,e)=>y*e+30+h)).attr("y",-5).attr("text-anchor","middle").attr("font-size",11).text((t=>`${t.upperBound}`))):(f.selectAll("rect").data(x).enter().append("rect").attr("fill",(t=>t.color)).attr("x",((t,e)=>y*e+h)).attr("width",y).attr("height",15).on("click",w),f.selectAll("text").data(x).join("text").attr("x",((t,e)=>y*e+30+h)).attr("y",-5).attr("text-anchor","middle").attr("font-size",11).text((t=>`${t.lowerBound.toFixed(0)} - ${t.upperBound.toFixed(0)}`)),f.append("text").attr("dy",28).attr("x",h).attr("font-size",14).text("Click on category to select/deselect days"))}function d(){const t=document.getElementById("button-text");switch(n){case 0:t.textContent="Mood Heatmap";break;case 1:t.textContent="Minutes of Sleep Heatmap";break;case 2:t.textContent="Minutes of Activity Heatmap";break;default:console.log("UHHHH")}}document.getElementById("uploadForm").addEventListener("submit",(function(t){t.preventDefault();const e=document.getElementById("myfile").files[0];if(e){const t=new FileReader;t.onload=function(t){var e;e=t.target.result,console.log("Process"),r=d3.csvParse(e),d(),l()},t.readAsText(e)}})),document.addEventListener("DOMContentLoaded",(function(){const t=document.querySelector(".arrow-left"),e=document.querySelector(".arrow-right");t.addEventListener("click",(function(){console.log("Left arrow button clicked"),n=(n+2)%3,d(),l()})),e.addEventListener("click",(function(){console.log("Right arrow button clicked"),n=(n+1)%3,d(),l()}))}))})();
